Menghani & Matani updates and random la operations added

+++ vec.c
@@ -23,7 +23,7 @@ int vec_expand_(char **data, int *length, int *capacity, int memsz) {
 int vec_reserve_(char **data, int *length, int *capacity, int memsz, int n) {
   (void) length;
   if (n > *capacity) {
-    void *ptr = realloc(*data, n * (long)memsz);
+    void *ptr = realloc(*data, n * memsz);
     if (ptr == NULL) return -1;
     *data = ptr;
     *capacity = n;
